Add an endpoint for updating a device and handle exceptions

+++ service/switchboardapp/src/main/java/com/switchboard/app/repository/EncoderRepository.java
@@ -14,6 +14,6 @@ public interface EncoderRepository extends JpaRepository<EncoderEntity,String> {
     EncoderEntity save(EncoderEntity decoder);
     Optional<EncoderEntity> findEncoderBySerialNumber(String serialNumber);
     List<EncoderEntity> findAll();
-    Long deleteEncoderEntityBySerialNumber(String serialNumber);
+    int deleteEncoderEntityBySerialNumber(String serialNumber);
 
 }
